updated other classes to work with grid and card refactor
