+++ src/main/java/sid/roborally/game_mechanics/grid/Grid.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 import java.util.HashSet;
 
 public class Grid {
-    ArrayList<ArrayList<HashSet<GridObject>>> grid;
+    public ArrayList<ArrayList<HashSet<GridObject>>> grid;
 
     private int width;
     private int height;
