+++ src/test/java/sid.roborally/game_mechanics/CardTest.java
@@ -1,6 +1,7 @@
 package sid.roborally.game_mechanics;
 
 import org.junit.Test;
+import sid.roborally.game_mechanics.card.Card;
 
 import static org.junit.Assert.assertEquals;
