+++ src/test/java/sid.roborally/game_mechanics/GridObjectTest.java
@@ -2,6 +2,8 @@ package sid.roborally.game_mechanics;
 
 import org.junit.Before;
 import org.junit.Test;
+import sid.roborally.game_mechanics.grid.GridObject;
+import sid.roborally.game_mechanics.grid.Position;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
