+++ src/test/java/sid.roborally/game_mechanics/GridTest.java
@@ -2,6 +2,10 @@ package sid.roborally.game_mechanics;
 
 import org.junit.Before;
 import org.junit.Test;
+import sid.roborally.game_mechanics.grid.Grid;
+import sid.roborally.game_mechanics.grid.GridObject;
+import sid.roborally.game_mechanics.grid.Position;
+import sid.roborally.game_mechanics.grid.Robot;
 
 import java.util.ArrayList;
 import java.util.HashSet;
