+++ src/test/java/sid.roborally/game_mechanics/HoleTest.java
@@ -1,6 +1,8 @@
 package sid.roborally.game_mechanics;
 
 import org.junit.Test;
+import sid.roborally.game_mechanics.grid.Hole;
+import sid.roborally.game_mechanics.grid.Position;
 
 import static org.junit.Assert.assertTrue;
