+++ src/test/java/sid.roborally/game_mechanics/PositionTest.java
@@ -1,6 +1,7 @@
 package sid.roborally.game_mechanics;
 
 import org.junit.Test;
+import sid.roborally.game_mechanics.grid.Position;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
