+++ src/test/java/sid.roborally/game_mechanics/StepCardTest.java
@@ -1,7 +1,6 @@
 package sid.roborally.game_mechanics;
-import org.hamcrest.Condition;
-import org.junit.Before;
 import org.junit.Test;
+import sid.roborally.game_mechanics.card.StepCard;
 
 import static org.junit.Assert.assertEquals;
 public class StepCardTest {
